Deleting old files, moving uninplemented function
