--- frontend/js/cart.js
@@ -1,163 +1,0 @@
-let carrouselSalesInfo = [
-    {
-        "itemsPerPage": 5,
-        "title": "Interesses",
-        "items": [
-            { "curQtty": 3, "milestones": [{ "quantity": 0, "price": 50 }, { "quantity": 50, "price": 40 }, { "quantity": 75, "price": 25 }, { "quantity": 100, "price": 15 }], "alt": "item0", "id": "aid1c0", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "aid1c1", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/300" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "aid1c2", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/400" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "aid1c3", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/500" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "aid1c4", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/600" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "aid1c5", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/700" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "aid1c6", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/800" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "aid1c7", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/900" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "aid1c8", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/1000" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "aid1c9", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/1100" }
-
-        ]
-    },
-
-    {
-        "itemsPerPage": 2,
-        "title": "Para você",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item0", "id": "aid2c0", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "aid2c1", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "aid2c2", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "aid2c3", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "aid2c4", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "aid2c5", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "aid2c6", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "aid2c7", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "aid2c8", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "aid2c9", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-    {
-        "itemsPerPage": 6,
-        "title": "Próximos de esgotar",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 30, "price": 25 }, { "quantity": 500, "price": 15 }], "alt": "item0", "id": "aid3c0", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "aid3c1", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "aid3c2", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "aid3c3", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "aid3c4", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "aid3c5", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "aid3c6", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "aid3c7", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "aid3c8", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "aid3c9", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-    {
-        "itemsPerPage": 1,
-        "title": "Em promoção",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item0", "id": "aid4c0", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "aid4c1", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "aid4c2", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "aid4c3", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "aid4c4", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "aid4c5", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "aid4c6", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "aid4c7", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "aid4c8", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "aid4c9", "link": "/pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-];
-
-function getProductData(product, qttyInCartItem) {
-    let curItems = parseInt(product.curQtty) + qttyInCartItem;
-    let maxNearest = { quantity: -1, price: undefined };
-    let nextMaxNearest = { quantity: 99999999999999, price: undefined };
-    for (let milestone of product.milestones) {
-        if (milestone.quantity > maxNearest.quantity && milestone.quantity <= curItems) {
-            maxNearest = milestone;
-        }
-    }
-    
-    
-    for (let milestone of product.milestones) {
-        if ((milestone.quantity < nextMaxNearest.quantity || nextMaxNearest === undefined) && milestone.quantity > maxNearest.quantity) {
-            nextMaxNearest = milestone;
-        }
-    }
-
-    if (maxNearest.price === undefined) maxNearest.price = '-';
-    if (nextMaxNearest.price === undefined) nextMaxNearest = {...maxNearest};
-
-    let currPricePerItem = maxNearest.price;
-    let nextPricePerItem = nextMaxNearest.price;
-    let remainingToReducePrice = ( nextMaxNearest.quantity !== maxNearest.quantity ? nextMaxNearest.quantity - curItems : 0);
-    return {
-        currPricePerItem,
-        nextPricePerItem,
-        remainingToReducePrice,
-    };
-}
-
-
-$(document).ready(function () {
-
-    $('.quantity-right-plus').click(function (e) {
-        e.preventDefault();
-        updateQuantityDelta(this, +1)
-    });
-
-    $('.quantity-left-minus').click(function (e) {
-        e.preventDefault();
-        updateQuantityDelta(this, -1)
-    });
-
-    $('input.quantity').on('keypress', function(e){
-        e.preventDefault();
-        if (e.key < '0' || e.key > '9') return
-        updateQuantity(this, Math.max(0, Math.min(parseInt($(this).val()+e.key), 10000)));
-    })
-
-    function updateQuantityDelta(caller, delta) {
-        let cardParent = $(caller).parent().parent().parent().parent().parent();
-        let quantityField = cardParent.find('.quantity');
-        let newQuant = Math.max(0, parseInt(quantityField.val()) + delta);
-        updateQuantity(caller, newQuant);
-    }
-
-    function updateQuantity(caller, newQuant) {
-        let cardParent = $(caller).parent().parent().parent().parent().parent();
-        let quantityField = cardParent.find('.quantity');
-        quantityField.val(newQuant);
-        cardParent.find('span.quantity .value').html(newQuant);
-
-        let productData = getProductData(carrouselSalesInfo[0].items[0], newQuant);
-
-        let { 
-            currPricePerItem,
-            nextPricePerItem,
-            remainingToReducePrice 
-        } = productData;
-
-        cardParent.find('span.cart-item-price .value').html(currPricePerItem);
-        cardParent.find('.cart-item-total .value').html(newQuant * currPricePerItem);
-
-        cardParent.find('span.cart-item-remaining-to-decrease .value').html(remainingToReducePrice);
-        cardParent.find('span.cart-item-next-value .value').html(nextPricePerItem);
-        cardParent.find('span.cart-item-min-next-value .value').html((newQuant + remainingToReducePrice)*nextPricePerItem);
-
-
-
-        updateTotal();
-    }
-
-    function updateTotal() {
-        let total = 0;
-        $('.cart-item-total .value').each(function (_) { total += parseInt($(this).html()) });
-        $('.cart-total .value').html(total);
-    }
-
-    $('input.quantity').each(function(_) { updateQuantity(this, 1)});
-
-});
