--- frontend/js/index.js
@@ -1,338 +1,0 @@
-let layerInfo = {
-    "layers": {
-        "1": [
-            { "name": "item0", "id": "id0", "parent": "undefined", "imageSrc": "./img/fridge.png" },
-            { "name": "item1", "id": "id1", "parent": "undefined", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item2", "id": "id2", "parent": "undefined", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item3", "id": "id3", "parent": "undefined", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item4", "id": "id4", "parent": "undefined", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item5", "id": "id5", "parent": "undefined", "imageSrc": "https://picsum.photos/200" }/*,
-            { "name": "item6", "id": "id6", "parent": "undefined", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item7", "id": "id7", "parent": "undefined", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item8", "id": "id8", "parent": "undefined", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item9", "id": "id9", "parent": "undefined", "imageSrc": "https://picsum.photos/200" }*/
-        ],
-        "11": [
-            { "name": "item10a", "id": "id10a", "parent": "id0", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item10b", "id": "id10b", "parent": "id0", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item10c", "id": "id10c", "parent": "id0", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item10d", "id": "id10d", "parent": "id0", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item10e", "id": "id10e", "parent": "id0", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item11", "id": "id11", "parent": "id1", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item12", "id": "id12", "parent": "id2", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item13", "id": "id13", "parent": "id3", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item14", "id": "id14", "parent": "id4", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item15", "id": "id15", "parent": "id5", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item16", "id": "id16", "parent": "id6", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item17", "id": "id17", "parent": "id7", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item18", "id": "id18", "parent": "id8", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item19", "id": "id19", "parent": "id9", "imageSrc": "https://picsum.photos/200" }
-
-        ],
-        "111": [
-            { "name": "item20", "id": "id20", "parent": "id10a", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item21", "id": "id21", "parent": "id11", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item22", "id": "id22", "parent": "id12", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item23", "id": "id23", "parent": "id13", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item24", "id": "id24", "parent": "id14", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item25", "id": "id25", "parent": "id15", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item26", "id": "id26", "parent": "id16", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item27", "id": "id27", "parent": "id17", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item28", "id": "id28", "parent": "id18", "imageSrc": "https://picsum.photos/200" },
-            { "name": "item29", "id": "id29", "parent": "id19", "imageSrc": "https://picsum.photos/200" }
-        ]
-    }
-};
-
-let carrouselLayerInfo = {
-    "itemsPerPage": 5,
-    "items": [
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item0", "id": "id0", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2100" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "id1", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2200" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "id2", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2300" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "id3", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2400" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "id4", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2500" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "id5", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2600" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "id6", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2700" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "id7", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2000" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "id8", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2900" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "id9", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2200" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item10", "id": "id10", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3000" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item11", "id": "id11", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3100" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item12", "id": "id12", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3200" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item13", "id": "id13", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3200" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item14", "id": "id14", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3400" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item15", "id": "id15", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3000" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item16", "id": "id16", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3600" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item17", "id": "id17", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/2700" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item18", "id": "id18", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3800" },
-        { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item19", "id": "id19", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/3900" }
-
-    ]
-};
-
-let carrouselSalesInfo = [
-    {
-        "itemsPerPage": 5,
-        "title": "Interesses",
-        "items": [
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 6, "price": 15 }], "alt": "item0", "id": "id1c0", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "id1c1", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/300" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "id1c2", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/400" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "id1c3", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/500" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "id1c4", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/600" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "id1c5", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/700" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "id1c6", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/800" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "id1c7", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/900" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "id1c8", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/1000" },
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "id1c9", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/1100" }
-
-        ]
-    },
-
-    {
-        "itemsPerPage": 2,
-        "title": "Para você",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item0", "id": "id2c0", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "id2c1", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "id2c2", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "id2c3", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "id2c4", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "id2c5", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "id2c6", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "id2c7", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "id2c8", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "id2c9", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-    {
-        "itemsPerPage": 6,
-        "title": "Próximos de esgotar",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 30, "price": 25 }, { "quantity": 500, "price": 15 }], "alt": "item0", "id": "id3c0", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "id3c1", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "id3c2", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "id3c3", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "id3c4", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "id3c5", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "id3c6", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "id3c7", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "id3c8", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "id3c9", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-    {
-        "itemsPerPage": 1,
-        "title": "Em promoção",
-        "items": [
-            { "curQtty": 4, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item0", "id": "id4c0", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item1", "id": "id4c1", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 1, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item2", "id": "id4c2", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 9, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item3", "id": "id4c3", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 7, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item4", "id": "id4c4", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item5", "id": "id4c5", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 3, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item6", "id": "id4c6", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 2, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item7", "id": "id4c7", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 5, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item8", "id": "id4c8", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" },
-            { "curQtty": 0, "milestones": [{ "quantity": 3, "price": 25 }, { "quantity": 5, "price": 15 }], "alt": "item9", "id": "id4c9", "link": "./pages/buypage.html", "imageSrc": "https://picsum.photos/200" }
-
-        ]
-    },
-];
-
-function getItemAttrs(item) {
-    let minPriceMile = { price: item.milestones[0].price + 1 };
-    let curPrice;
-    let curQtty = item.curQtty;
-    let nearestMile = { "quantity": -1 };
-    let biggestMile = { "quantity": -1 };
-    for (milestone of item.milestones) {
-        if (milestone.quantity > biggestMile.quantity)
-            biggestMile.quantity = milestone.quantity;
-
-        if (curQtty > milestone.quantity && milestone.quantity > nearestMile.quantity)
-            nearestMile = milestone;
-
-        if (milestone.price < minPriceMile.price) minPriceMile = milestone;
-    }
-
-    if (biggestMile.quantity === -1) {
-        console.error('Invalid product: no valid milestones. ID = ', item.id);
-    }
-
-    if (nearestMile.quantity === -1) curPrice = '-';
-    else curPrice = nearestMile.price;
-
-    return { curQtty, maxQuantity: biggestMile.quantity, curPrice, minPrice: minPriceMile.price };
-}
-
-function generateItem(item) {
-    let itemAttrs = getItemAttrs(item);
-    return `
-        <div class="col mx-auto" style="min-width:00px; max-width:200px">
-            <div class="card">
-                <a href=${item.link}>
-                    <img src="${item.imageSrc}" class="d-block img-thumbnail mx-auto" alt="${item.alt}" />
-                </a>
-                <div class="card-body px-1 pt-2">
-                    <div class="row g-0 p-0">
-                        <div class="col-10" style="text-transform:capitalize;"> ${item.alt} </div>
-                    </div>
-                    <div class="row flex-direction-row mt-2 px-1">
-                        <div class="col-6">
-                            R$${itemAttrs.curPrice}
-                        </div>
-                        <div class="col-6 text-end">
-                            R$${itemAttrs.minPrice}
-                        </div>
-                    </div>
-                    <div class="progress">
-                        <div class="progress-bar text-center bg-warning" role="progressbar" style="width: ${100 * itemAttrs.curQtty / itemAttrs.maxQuantity}%" aria-valuenow="10" aria-valuemin="0" aria-valuemax="100">
-                        <small class="justify-content-center d-flex position-absolute w-100 text-dark fw-bold">${itemAttrs.curQtty}/${itemAttrs.maxQuantity}</small>
-                        
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    `
-}
-
-
-$(document).ready(function () {
-    async function populateCarrousel(carrouselType, location, idName, generateItem, classes) {
-        classes = classes || '';
-
-        let itemsPerPage = Math.min(carrouselType.itemsPerPage, carrouselType.items.length);
-        if (itemsPerPage <= 0) return;
-
-        let newHtml = `<div class="row layer ${classes} carousel-container mt-2" id="carousel-${idName}-row">
-                        <h4 class="${classes || 'mt-5'} w-100 text-center ${classes}  ">${carrouselType.title}</h4>
-                        <div class="col-9 mx-auto w-100">
-                            <div class="carousel slide" id="carousel-${idName}" data-bs-ride="carousel">
-                                <div class="carousel-inner w-75 mx-auto">`;
-
-
-        let pageCount = Math.ceil(carrouselType.items.length / itemsPerPage);
-        let lastPageItemsToAdd = (itemsPerPage - carrouselType.items.length % itemsPerPage) % itemsPerPage;
-        for (let i = 0; i < pageCount; i++) {
-            newHtml += `<div class="${'carousel-item row-car ' + (i === 0 ? 'active' : '')} ">
-                            <div class="row g-1">`
-            let k = i * itemsPerPage;
-
-            if (i === pageCount - 1) k -= lastPageItemsToAdd;
-
-            for (let j = 0; j < itemsPerPage; j++) {
-                let item = carrouselType.items[k + j];
-                newHtml += generateItem(item);
-            }
-            newHtml += `
-                            </div>
-                        </div>
-            `;
-        }
-        newHtml += "</div>";
-        newHtml += `
-                    <button class="carousel-control-prev" type="button" data-bs-target="#carousel-${idName}" data-bs-slide="prev">
-                        <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-                        <span class="visually-hidden">Previous</span>
-                    </button>
-                    <button class="carousel-control-next" type="button" data-bs-target="#carousel-${idName}" data-bs-slide="next">
-                        <span class="carousel-control-next-icon" aria-hidden="true"></span>
-                        <span class="visually-hidden">Next</span>
-                    </button>
-        `
-
-        newHtml += `
-                        </div>
-                    </div>
-                </div>
-            `;
-
-        let currHtml = $(location).html();
-        $(location).html(currHtml + newHtml);
-
-    }
-
-    function populateLayers() {
-        let newHtml = ''
-
-        for (let layer in layerInfo.layers) {
-            newHtml += `
-                <div class="row layer ${layer !== '1' ? 'd-none' : ''}" data-layer="${layer}">
-                    <div class="mx-auto align-self-center d-flex flex-row ">
-            `;
-            for (let category of layerInfo.layers[layer]) {
-                newHtml += `
-                <div class="category mx-auto ${layer !== '1' ? 'd-none' : ''}" id="${category.id}" data-parent="${category.parent}" style="background:url(${category.imageSrc});">
-                    ${category.name}
-                </div>`;
-            }
-
-            newHtml += "</div>\n</div>"
-
-        }
-
-
-        let currHtml = $('#categories').html();
-        $('#categories').html(currHtml + newHtml);
-
-
-
-
-        let cats = $(".category");
-
-        for (let cat of cats) {
-            let id = $(cat).attr('data-parent');
-            $('.category#' + id).addClass('parent');
-        }
-
-    }
-
-    populateLayers();
-    populateCarrousel(carrouselLayerInfo, '#categories', 'layer', generateItem, 'd-none');
-
-    for (let i = 0; i < carrouselSalesInfo.length; i++) populateCarrousel(carrouselSalesInfo[i], '#sales', `sale-${i}`, generateItem);
-
-
-    //Mouse Events
-    $('.category').mouseover(function (e) {
-        e.preventDefault();
-        let self = $(this);
-        let selfLayer = $(this).parent().parent().attr('data-layer');
-        let targetRemoveLayerNum = selfLayer + '1';
-
-        let targetRemoveLayer = $(`#categories .layer[data-layer^=${targetRemoveLayerNum}]`);
-        targetRemoveLayer.addClass('d-none');
-        targetRemoveLayer.find('.category').addClass('d-none');
-
-
-        let targets = $('#categories [data-parent="' + self.attr('id') + '"]');
-        let targetLayers = targets.parent().parent();
-        targetLayers.removeClass('d-none');
-        targets.removeClass('d-none');
-        console.log(self.hasClass('parent'))
-
-        if (!self.hasClass('parent')) {
-            $('#carousel-layer-row').removeClass('d-none');
-        }
-        else {
-            $('#carousel-layer-row').addClass('d-none');
-        }
-
-    });
-
-    $('#categories').mouseleave(function (e) {
-        e.preventDefault();
-
-        let layers = $(this).find(".layer[data-layer]:not([data-layer='1']")
-        layers.addClass('d-none');
-        layers.find('.category').addClass('d-none');
-        $('#carousel-layer-row').addClass('d-none');
-
-
-    });
-});
